# gaggiuino-local-profiler/openapi.yaml
openapi: 3.0.3
info:
  title: GLP REST API
  version: 2.0.0
  description: >
    REST API for the **Gaggiuino Local Profiler** Home Assistant add-on.
    Provides shot storage, coffee library, maintenance tracking, a barista
    orders system, and live machine data.
  contact:
    url: https://github.com/mxkissnr/gaggiuino-local-profiler
servers:
  - url: http://localhost:8099
    description: Direct port (integration, cards)
  - url: /api/hassio_ingress/gaggiuino_local_profiler
    description: HA Ingress (browser — auth via HA session, no token needed)
security:
  - tokenAuth: []
components:
  securitySchemes:
    tokenAuth:
      type: apiKey
      in: header
      name: X-GLP-Token
      description: >
        64-char hex token, auto-generated on first start, stored in
        `/data/api_token.txt`. Retrieve from `GET /api/status`.
        Requests via HA Ingress bypass this check (HA session is sufficient).
        `GET /api/status` is always public.
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, example: "not found" }
...
    MaintenanceTask:
      type: object
      properties:
        lastDate:        { type: string, format: date, nullable: true }
        threshold_shots: { type: integer, minimum: 1, maximum: 10000, nullable: true }
        threshold_days:  { type: integer, minimum: 1, maximum: 365, nullable: true }
        shots_since:     { type: integer, description: Shots since last maintenance }
        days_since:      { type: integer, description: Days since last maintenance, nullable: true }
        pct:             { type: number, minimum: 0, maximum: 1, description: Progress toward threshold (0–1) }
        status:          { type: string, enum: [ok, warning, overdue] }
    MaintenanceStats:
      type: object
      description: >
        Keys are task names: `descaling`, `backflush`, `grouphead`,
        `gaskets`, `waterfilter`, `grinder_{id}`.
      additionalProperties:
        $ref: '#/components/schemas/MaintenanceTask'

    # ── Orders ────────────────────────────────────────────────────────────
...
paths:
...
  /api/maintenance/{task}/done:
    post:
      tags: [Maintenance]
      summary: Mark a maintenance task as done today
      parameters:
        - name: task
          in: path
          required: true
          schema: { type: string }
          description: Task name (e.g. "descaling", "grinder_1716000000000")
      responses:
        "200": { description: Updated stats, content: { application/json: { schema: { $ref: '#/components/schemas/MaintenanceStats' } } } }
        "404": { description: Unknown task, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/maintenance/{task}/threshold:
    post:
      tags: [Maintenance]
      summary: Update reminder thresholds for a maintenance task
      parameters:
        - name: task
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                threshold_shots: { type: integer, minimum: 1, maximum: 10000, nullable: true }
                threshold_days:  { type: integer, minimum: 1, maximum: 365, nullable: true }
      responses:
        "200": { description: Updated stats, content: { application/json: { schema: { $ref: '#/components/schemas/MaintenanceStats' } } } }
        "404": { description: Unknown task, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }

  # ── Orders ───────────────────────────────────────────────────────────────
...